SellerInfo type
