+++ linx/utils/types/productList.ts
@@ -1,4 +1,5 @@
 import { LinxMetadata } from "./shared-types/linxMetada.ts";
+import { SellerInfo } from "./shared-types/sellerInfo.ts";
 
 export interface ProductListResponse {
   HasPreviousPage: boolean;
@@ -267,15 +268,11 @@ export interface Price {
   SalesPriceWithTaxDiscount: number;
   TaxationAmount: number;
   HasBuyBox: boolean;
-  BuyBox: BuyBox;
+  BuyBox: SellerInfo;
   CurrentPaymentTerm: null;
   IsNullPrice: boolean;
 }
 
-export interface BuyBox {
-  SellerID: number;
-  SellerName: null;
-}
 
 export interface MaxInstallmentsNoInterest {
   Installments: number;
