+++ linx/utils/types/shared-types/categoryItem.ts
@@ -1,5 +1,5 @@
 export interface CategoryItem {
-    ID: string;
-    Name: string;
-    ParentID: string;
-  }
+  ID: string;
+  Name: string;
+  ParentID: string;
+}
