+++ linx/utils/types/shared-types/navigationInfo.ts
