--- linx/utils/types/shared-types/productSelection.ts
@@ -1,6 +1,0 @@
-export interface ProductSelectionItem {
-    PropertyLevel: number;
-    PropertyPath: string;
-    Quantity: number;
-    SkuID: number;
-  }
