+++ linx/utils/types/shared-types/sellerInfo.ts
@@ -1,0 +1,5 @@
+export interface SellerInfo {
+    SellerID: number;
+    SellerName: null;
+  }
+
