+++ linx/utils/types/suggestionsJSON.ts
@@ -1,9 +1,10 @@
 import { LinxError } from "./common.ts";
 import { CategoryItem } from "./shared-types/categoryItem.ts";
 import { LinxMetadata } from "./shared-types/linxMetada.ts";
-import { NavigationInfo } from "./shared-types/navigations.ts";
-import { ProductSelectionItem } from "./shared-types/productSelection.ts";
+import { NavigationInfo } from "./shared-types/navigationInfo.ts";
+import { ProductSelectionItem } from "./shared-types/productSelectionItem.ts";
 import { ResponseCallBack } from "./shared-types/responseCallBack.ts";
+import { SellerInfo } from "./shared-types/sellerInfo.ts";
 
 export interface WebPage {
   Response: Response;
@@ -154,7 +155,6 @@ export interface BrowsingImage {
   Width: number;
 }
 
-
 export interface Item {
   Availability: string;
   AvailabilityText: string;
@@ -268,7 +268,7 @@ export interface MediaGroup {
 export interface Price {
   BestInstallment: null;
   BestInstallmentCreditCard: null;
-  BuyBox: BuyBox;
+  BuyBox: SellerInfo;
   CurrentPaymentTerm: null;
   HasBuyBox: boolean;
   HasMunknownSkus: boolean;
@@ -285,11 +285,6 @@ export interface Price {
   SalesPriceWithTaxDiscount: number;
   Tax: number;
   TaxationAmount: number;
-}
-
-export interface BuyBox {
-  SellerID: number;
-  SellerName: null;
 }
 
 export interface MaxInstallmentsNoInterest {
